+++ site/views/app/view.html.php
@@ -17,10 +17,10 @@
 
 // No direct access to this file
 defined('_JEXEC') or die('Restricted access');
-jimport('joomla.application.module.helper');
+use Joomla\CMS\Helper\ModuleHelper as JModuleHelper;
+use VDM\Joomla\GetBible\Factory;
 
 use Joomla\CMS\MVC\View\HtmlView;
-use VDM\Joomla\GetBible\Factory;
 
 /**
  * Getbible Html View class for the App
@@ -929,7 +929,6 @@ class GetbibleViewApp extends HtmlView
 				default:
 					return implode('<br />', $this->setModules[$position]);
 					break;
-				
 			}
 		}
 		return false;